ORS-8: Redux Components
