+++ src/app.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { createStore } from 'redux'
+import { Provider } from 'react-redux';
 
 import { ThemeProvider } from '@material-ui/styles';
 
@@ -9,22 +11,25 @@ import Product from './pages/product';
 import Products from './pages/products';
 import Footer from './components/footer';
 
+import MainReducer from './store/reducer';
 import AppTheme from './app-theme';
 
 
 const App: React.FC = () => {
   return (
-    <Router>
-      <ThemeProvider theme={AppTheme}>
-        <NavBar/>
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route exact path="/products" component={Products} />
-            <Route exact path="/products/:id" component={Product} />
-          </Switch>
-        <Footer/>
-      </ThemeProvider>
-    </Router>
+    <Provider store={createStore(MainReducer)}>
+      <Router>
+        <ThemeProvider theme={AppTheme}>
+          <NavBar/>
+            <Switch>
+              <Route exact path="/" component={Home} />
+              <Route exact path="/products" component={Products} />
+              <Route exact path="/products/:id" component={Product} />
+            </Switch>
+          <Footer/>
+        </ThemeProvider>
+      </Router>
+    </Provider>
   );
 }
