+++ src/common/reducer.tsx
@@ -1,4 +1,15 @@
-import { IStore, IAction, IHandlers } from '../types';
+export interface IStore {
+  
+}
+
+export interface IAction {
+  type: string,
+  payload: any,
+}
+
+export interface IHandlers {
+  [key: string]: any
+}
 
 export default (initialState: IStore, handlers: IHandlers) => {
   return function reducer(state = initialState, action: IAction) {
