--- src/store/types.tsx
@@ -1,12 +1,0 @@
-export interface IStore {
-  
-}
-
-export interface IAction {
-  type: string,
-  payload: any,
-}
-
-export interface IHandlers {
-  [key: string]: any
-}
